공지 사항 API 변경된 부분 반영 및 클릭 이벤트 구현  (#195)
* 공지 API endpoint 변경

* 홈 화면 공지사항 목록 recyclerview 고정 높이 적용

* 공지 사항 DTO 수정 및, 시안에 맞게 날짜 표기 방식 변경

* id Data type Int -> Long 으로 변경, 홈 화면 공지 카드 눌렀을 때 공지화면으로 넘어가도록 ClickListener 구현

int64 는 Long

* spotless check success

+++ common/src/main/kotlin/us/wedemy/eggeum/android/common/model/FileModel.kt
@@ -12,6 +12,6 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class FileModel(
-  val uploadFileId: Int?,
+  val uploadFileId: Long?,
   val url: String?,
 ) : Parcelable
