+++ domain/src/main/kotlin/us/wedemy/eggeum/android/domain/model/file/FileResponseEntity.kt
@@ -9,6 +9,6 @@ package us.wedemy.eggeum.android.domain.model.file
 
 public data class FileResponseEntity(
   val name: String?,
-  val uploadFileId: Int?,
+  val uploadFileId: Long?,
   val url: String?,
 )
